+++ db/data_schema.rb
@@ -1,2 +1,2 @@
 # encoding: UTF-8
-DataMigrate::Data.define(version: 20191203100302)
+DataMigrate::Data.define(version: 20191216151537)