+++ db/migrate/20191216135816_add_checked_by_reference_to_checks.rb
@@ -1,0 +1,5 @@
+class AddCheckedByReferenceToChecks < ActiveRecord::Migration[6.0]
+  def change
+    add_reference :checks, :checked_by, type: :uuid, foreign_key: {to_table: :dfe_sign_in_users}
+  end
+end