+++ spec/requests/admin_payroll_run_downloads_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe "Admin payroll run downloads" do
   let(:admin_session_id) { "some_user_id" }
 
   before do
-    sign_in_to_admin_with_role(AdminSession::PAYROLL_OPERATOR_DFE_SIGN_IN_ROLE_CODE, admin_session_id)
+    sign_in_to_admin_with_role(DfeSignIn::User::PAYROLL_OPERATOR_DFE_SIGN_IN_ROLE_CODE, admin_session_id)
   end
 
   describe "downloads#new" do
@@ -116,7 +116,7 @@ RSpec.describe "Admin payroll run downloads" do
   end
 
   describe "When signed in as a service operator or a support agent, download routes" do
-    [AdminSession::SERVICE_OPERATOR_DFE_SIGN_IN_ROLE_CODE, AdminSession::SUPPORT_AGENT_DFE_SIGN_IN_ROLE_CODE].each do |role|
+    [DfeSignIn::User::SERVICE_OPERATOR_DFE_SIGN_IN_ROLE_CODE, DfeSignIn::User::SUPPORT_AGENT_DFE_SIGN_IN_ROLE_CODE].each do |role|
       it "respond with not authorised" do
         payroll_run = create(:payroll_run)